u-boot: Allow for customizing installed/deployed file names
When assembling all of the various filenames that are installed/deployed
from u-boot, we have been including the PV and PR in the filenames.  This
change introduces a single variable to replace these two in the
filenames.

This change should not be disruptive since the default value for the new
UBOOT_VERSION variable is "${PV}-${PR}".

(From OE-Core rev: 1cc3db8909e71a5f8d5ab566c5e80344d9360489)

Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>